Overwrite branch to exactly match mateyu

+++ app/src/main/java/org/dhis2/usescases/main/MainPresenter.kt
@@ -3,15 +3,18 @@ package org.dhis2.usescases.main
 import android.content.Context
 import android.net.Uri
 import android.view.Gravity
+import androidx.core.net.toUri
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.asLiveData
 import androidx.work.ExistingWorkPolicy
-import io.reactivex.Completable
 import io.reactivex.Flowable
 import io.reactivex.disposables.CompositeDisposable
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
+import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.onEach
+import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.launch
 import org.dhis2.BuildConfig
 import org.dhis2.commons.Constants
@@ -26,7 +29,6 @@ import org.dhis2.commons.matomo.Labels.Companion.CLICK
 import org.dhis2.commons.matomo.MatomoAnalyticsController
 import org.dhis2.commons.prefs.Preference
 import org.dhis2.commons.prefs.Preference.Companion.DEFAULT_CAT_COMBO
-import org.dhis2.commons.prefs.Preference.Companion.PIN
 import org.dhis2.commons.prefs.Preference.Companion.PREF_DEFAULT_CAT_OPTION_COMBO
 import org.dhis2.commons.prefs.PreferenceProvider
 import org.dhis2.commons.schedulers.SchedulerProvider
@@ -39,6 +41,7 @@ import org.dhis2.data.service.workManager.WorkManagerController
 import org.dhis2.data.service.workManager.WorkerItem
 import org.dhis2.data.service.workManager.WorkerType
 import org.dhis2.usescases.login.SyncIsPerformedInteractor
+import org.dhis2.usescases.main.domain.LogoutUser
 import org.dhis2.usescases.settings.DeleteUserData
 import org.dhis2.usescases.sync.WAS_INITIAL_SYNC_DONE
 import org.dhis2.utils.TRUE
@@ -46,6 +49,8 @@ import org.hisp.dhis.android.core.organisationunit.OrganisationUnit
 import org.hisp.dhis.android.core.systeminfo.SystemInfo
 import org.hisp.dhis.android.core.user.User
 import timber.log.Timber
+import kotlin.concurrent.atomics.AtomicBoolean
+import kotlin.concurrent.atomics.ExperimentalAtomicApi
 import kotlin.coroutines.CoroutineContext
 
 const val DEFAULT = "default"
@@ -53,6 +58,7 @@ const val SERVER_ACTION = "Server"
 const val DHIS2 = "dhis2_server"
 const val PLAY_FLAVOR = "dhisPlayServices"
 
+@OptIn(ExperimentalAtomicApi::class)
 class MainPresenter(
     private val view: MainView,
     private val repository: HomeRepository,
@@ -67,10 +73,10 @@ class MainPresenter(
     private val syncIsPerformedInteractor: SyncIsPerformedInteractor,
     private val syncStatusController: SyncStatusController,
     private val versionRepository: VersionRepository,
-    private val dispatcherProvider: DispatcherProvider,
+    val dispatcherProvider: DispatcherProvider,
     private val forceToNotSynced: Boolean,
+    private val logoutUser: LogoutUser,
 ) : CoroutineScope {
-
     private var job = Job()
     override val coroutineContext: CoroutineContext
         get() = job + dispatcherProvider.io()
@@ -80,10 +86,21 @@ class MainPresenter(
     val versionToUpdate = versionRepository.newAppVersion.asLiveData(coroutineContext)
     val downloadingVersion = MutableLiveData(false)
 
+    private val _singleProgramNavigationChannel = Channel<HomeItemData>()
+    val singleProgramNavigationChannel =
+        _singleProgramNavigationChannel
+            .receiveAsFlow()
+            .onEach {
+                singleProgramNavigationDone.store(true)
+            }
+
+    private var singleProgramNavigationDone = AtomicBoolean(false)
+
     fun init() {
         preferences.removeValue(Preference.CURRENT_ORG_UNIT)
         disposable.add(
-            repository.user()
+            repository
+                .user()
                 .map { username(it) }
                 .subscribeOn(schedulerProvider.io())
                 .observeOn(schedulerProvider.ui())
@@ -94,7 +111,8 @@ class MainPresenter(
         )
 
         disposable.add(
-            repository.defaultCatCombo()
+            repository
+                .defaultCatCombo()
                 .subscribeOn(schedulerProvider.io())
                 .subscribe(
                     { categoryCombo ->
@@ -105,7 +123,8 @@ class MainPresenter(
         )
 
         disposable.add(
-            repository.defaultCatOptCombo()
+            repository
+                .defaultCatOptCombo()
                 .subscribeOn(schedulerProvider.io())
                 .subscribe(
                     { categoryOptionCombo ->
@@ -122,7 +141,8 @@ class MainPresenter(
 
     fun initFilters() {
         disposable.add(
-            Flowable.just(filterRepository.homeFilters())
+            Flowable
+                .just(filterRepository.homeFilters())
                 .subscribeOn(schedulerProvider.io())
                 .observeOn(schedulerProvider.ui())
                 .subscribe(
@@ -138,7 +158,8 @@ class MainPresenter(
         )
 
         disposable.add(
-            filterManager.asFlowable()
+            filterManager
+                .asFlowable()
                 .subscribeOn(schedulerProvider.io())
                 .observeOn(schedulerProvider.ui())
                 .subscribe(
@@ -158,7 +179,8 @@ class MainPresenter(
         )
 
         disposable.add(
-            filterManager.ouTreeFlowable()
+            filterManager
+                .ouTreeFlowable()
                 .subscribeOn(schedulerProvider.io())
                 .observeOn(schedulerProvider.ui())
                 .subscribe(
@@ -170,7 +192,8 @@ class MainPresenter(
 
     fun trackDhis2Server() {
         disposable.add(
-            repository.getServerVersion()
+            repository
+                .getServerVersion()
                 .subscribeOn(schedulerProvider.io())
                 .observeOn(schedulerProvider.ui())
                 .subscribe(
@@ -200,45 +223,42 @@ class MainPresenter(
         filterManager.addOrgUnits(selectedOrgUnits)
     }
 
-    private fun getUserUid(): String {
-        return try {
-            userManager.d2.userModule().user().blockingGet()?.uid() ?: ""
+    private fun getUserUid(): String =
+        try {
+            userManager.d2
+                .userModule()
+                .user()
+                .blockingGet()
+                ?.uid() ?: ""
         } catch (e: Exception) {
             ""
         }
-    }
 
     fun logOut() {
-        disposable.add(
-            Completable.fromCallable {
-                workManagerController.cancelAllWork()
-                syncStatusController.restore()
-                filterManager.clearAllFilters()
-                preferences.setValue(Preference.SESSION_LOCKED, false)
-                preferences.setValue(Preference.PIN_ENABLED, false)
-                userManager.d2.dataStoreModule().localDataStore().value(PIN).blockingDeleteIfExist()
-            }.andThen(
-                repository.logOut(),
+        launch(dispatcherProvider.ui()) {
+            logoutUser().fold(
+                onSuccess = { accounts ->
+                    view.goToLogin(accounts, isDeletion = false)
+                },
+                onFailure = {
+                    Timber.e(it)
+                },
             )
-                .subscribeOn(schedulerProvider.ui())
-                .observeOn(schedulerProvider.ui())
-                .subscribe(
-                    {
-                        view.goToLogin(repository.accountsCount(), isDeletion = false)
-                    },
-                    { Timber.e(it) },
-                ),
-        )
+        }
     }
 
     fun onDeleteAccount() {
         view.showProgressDeleteNotification()
         try {
+            repository.checkDeleteBiometricsPermission()
             workManagerController.cancelAllWork()
             syncStatusController.restore()
             deleteUserData.wipeCacheAndPreferences(view.obtainFileView())
             userManager.d2?.wipeModule()?.wipeEverything()
-            userManager.d2?.userModule()?.accountManager()?.deleteCurrentAccount()
+            userManager.d2
+                ?.userModule()
+                ?.accountManager()
+                ?.deleteCurrentAccount()
             view.cancelNotifications()
 
             view.goToLogin(repository.accountsCount(), isDeletion = true)
@@ -268,13 +288,12 @@ class MainPresenter(
         view.openDrawer(Gravity.START)
     }
 
-    private fun username(user: User): String {
-        return String.format(
+    private fun username(user: User): String =
+        String.format(
             "%s %s",
             if (user.firstName().isNullOrEmpty()) "" else user.firstName(),
             if (user.surname().isNullOrEmpty()) "" else user.surname(),
         )
-    }
 
     fun onNavigateBackToHome() {
         view.goToHome()
@@ -293,9 +312,7 @@ class MainPresenter(
             .syncDataForWorker(Constants.DATA_NOW, Constants.INITIAL_SYNC)
     }
 
-    fun observeDataSync(): StateFlow<SyncStatusData> {
-        return syncStatusController.observeDownloadProcess()
-    }
+    fun observeDataSync(): StateFlow<SyncStatusData> = syncStatusController.observeDownloadProcess()
 
     fun wasSyncAlreadyDone(): Boolean {
         if (forceToNotSynced) {
@@ -306,7 +323,10 @@ class MainPresenter(
 
     fun onDataSuccess() {
         launch(dispatcherProvider.io()) {
-            userManager.d2.dataStoreModule().localDataStore().value(WAS_INITIAL_SYNC_DONE)
+            userManager.d2
+                .dataStoreModule()
+                .localDataStore()
+                .value(WAS_INITIAL_SYNC_DONE)
                 .blockingSet(TRUE)
         }
     }
@@ -330,12 +350,13 @@ class MainPresenter(
     }
 
     fun remindLaterAlertNewVersion() {
-        val workerItem = WorkerItem(
-            Constants.NEW_APP_VERSION,
-            WorkerType.NEW_VERSION,
-            delayInSeconds = 24 * 60 * 60,
-            policy = ExistingWorkPolicy.REPLACE,
-        )
+        val workerItem =
+            WorkerItem(
+                Constants.NEW_APP_VERSION,
+                WorkerType.NEW_VERSION,
+                delayInSeconds = 24 * 60 * 60,
+                policy = ExistingWorkPolicy.REPLACE,
+            )
         workManagerController.beginUniqueWork(workerItem)
         versionRepository.removeVersionInfo()
     }
@@ -347,7 +368,7 @@ class MainPresenter(
     ) {
         if (BuildConfig.FLAVOR == PLAY_FLAVOR) {
             val url = versionRepository.getUrl()
-            onLaunchUrl(Uri.parse(url))
+            url?.toUri()?.let { onLaunchUrl(it) }
         } else {
             versionRepository.download(
                 context = context,
@@ -360,15 +381,21 @@ class MainPresenter(
         }
     }
 
-    fun hasOneHomeItem(): Boolean {
-        return repository.homeItemCount() == 1
+    fun checkSingleProgramNavigation() {
+        if (!singleProgramNavigationDone.load() && repository.homeItemCount() == 1) {
+            launch(coroutineContext) {
+                repository.singleHomeItemData()?.let {
+                    _singleProgramNavigationChannel.send(it)
+                }
+            }
+        }
     }
 
-    fun getSingleItemData(): HomeItemData? {
-        return repository.singleHomeItemData()
+    fun hasFilters(): Boolean = filterRepository.homeFilters().isNotEmpty()
+
+    fun updateSingleProgramNavigationDone(done: Boolean) {
+        singleProgramNavigationDone.store(done)
     }
 
-    fun hasFilters(): Boolean {
-        return filterRepository.homeFilters().isNotEmpty()
-    }
+    fun isSingleProgramNavigationDone() = singleProgramNavigationDone.load()
 }
